+++ odoo/local-src/qoqa_offer/__openerp__.py
@@ -33,7 +33,6 @@ Adds *Offer* models.
           'account_invoice_view.xml',
           'stock_view.xml',
           'qoqa_offer_view.xml',
-          'delivery_carrier_view.xml',
           'procurement_view.xml',
           'security/ir.model.access.csv',
           'res_users_view.xml',